few Tweaks
Created a template folder and add 2 files footer and header for the footer and header. I changed the icons for the app bar. Updated the Filters and added a new field called type with lost and found type. I Tweaked the searchbar code to work as you scroll to never dissapear.